+++ spec/dummy/db/migrate/20190226172951_create_user.rb
@@ -1,0 +1,9 @@
+class CreateUser < ActiveRecord::Migration
+  def change
+    create_table :users do |t|
+      t.string :name
+
+      t.timestamps
+    end
+  end
+end