+++ spec/dummy/db/migrate/20190226173051_create_isle.rb
@@ -1,0 +1,10 @@
+class CreateIsle < ActiveRecord::Migration
+  def change
+    create_table :isle do |t|
+      t.string :name
+      t.binary :map
+
+      t.timestamps
+    end
+  end
+end