+++ spec/dummy/db/migrate/20190226174951_create_tree.rb
@@ -1,0 +1,12 @@
+class CreateTree < ActiveRecord::Migration
+  def change
+    create_table :trees do |t|
+      t.string :name
+      t.references :owner, index: true
+      t.references :cutter, index: true
+      t.references :island, index: true
+
+      t.timestamps
+    end
+  end
+end